% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{calc_confint}
\alias{calc_confint}
\title{Calculate the confidence intervals}
\usage{
calc_confint(
  result,
  standata,
  model,
  pars,
  method = "HDI",
  level = 0.95,
  options = list()
)
}
\arguments{
\item{result}{The result of the optimised model.}

\item{standata}{The data passed to stan.}

\item{model}{The stan model.}

\item{pars}{The parameters that were optimised.}

\item{method}{The method for calculating the confidence interval.
One of 'quap', 'ETI', 'HDI', 'BCI', 'SI', see details.}

\item{level}{The confidence level to use when calculating confidence intervals}

\item{options}{Options for the sampler}
}
\description{
Calculate the confidence intervals
}
\details{
\code{quap}: Confidence interval is computed using quadratic approximation.
For ETI, HDI, BCI and SI, the confidence interval is computed across samples from the posterior distribution.
\code{ETI}: Equal-Tailed Interval. The probability of being below this interval is equal to the probability of being above it. A 95\% ETI will always have 2.5\% of the distribution on either side of its limits. See https://easystats.github.io/bayestestR/reference/eti.html
\code{HDI}: Highest Density Interval. Not equal-tailed. See https://easystats.github.io/bayestestR/reference/hdi.html
\code{BCI}: Bias Corrected and Accelerated Interval. See https://easystats.github.io/bayestestR/reference/bci.html
\code{SI}: Support Interval. 'A support interval contains only the values of the parameter that predict the observed data better than average'. See https://easystats.github.io/bayestestR/reference/si.html
}
