% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_titer_lims.R
\name{calc_titer_lims}
\alias{calc_titer_lims}
\title{Get titer limits}
\usage{
calc_titer_lims(titers, dilution_stepsize, options = list())
}
\arguments{
\item{titers}{A numeric/character vector or matrix of titer measurements.}

\item{dilution_stepsize}{The dilution stepsize, see details}

\item{options}{A named list of options to pass to \code{titer_fit_options()}}
}
\value{
Returns a list of length three with values log_titers, max_titers and min_titers, giving the
numeric vectors of the straight log converted, upper and lower bounds of the titers on the log scale.
}
\description{
Function for getting upper and lower limits of measured titers on the log scale.
}
\details{
HI titers which typically follow a 2-fold dilution series starting
from 1/10, then 1/20, 1/40 etc. This represents a "dilution stepsize" of 1
when converted to the log2 scale. When no inhibition was recorded at the
highest dilution, the value is typically recorded as <10 but the
optimization regime effectively treats this as a <=5, the rationale being
that, had the dilution series been continued to higher concentrations, the
next lowest titer would have been a 5. Over time the method has also been
applied to other neutralization assays that sometimes have a continuous
read out with a lower end, in these cases a <10 really means a <10 since
any other values like 9.8 or 7.62 would also be possible. To indicate these
continuous cases, you can specify the dilution stepsize as 0. Equally, if
the dilution regime followed a different pattern, you can also set that
here.
}
\examples{
# Calculate the titer limits of a set of HI titers
titer_lims <- calc_titer_lims(
  titers = c("20", "320", "<10", ">1280"),
  dilution_stepsize = 1
)

# Calculate the titer limits assuming non-default upper and lower bounds for non-detectable
# and greater-than titers.
titer_lims <- calc_titer_lims(
  titers = c("20", "320", "<10", ">1280"),
  dilution_stepsize = 1,
  options = list(
    min_titer_possible = -Inf,
    max_titer_possible = 14
  )
)

}
