% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_crossreactivity_breadth.R
\name{compare_crossreactivity_breadth}
\alias{compare_crossreactivity_breadth}
\title{Compare cross-reactivity breadth}
\usage{
compare_crossreactivity_breadth(
  titers,
  sr_groups,
  homologous_ag,
  reference_sr_group,
  dilution_stepsize
)
}
\arguments{
\item{titers}{A named matrix of titer data, where column names are the serum
names and row names are the antigen names. Censored data should be included
as e.g. "<20"}

\item{sr_groups}{A character vector of serum group names for each of the sera
in the titer table}

\item{homologous_ag}{A string giving the name of the homologous antigen for
all the serum groups (should be only 1 antigen name from those included in
the table)}

\item{reference_sr_group}{A string giving the name of the reference serum group}

\item{dilution_stepsize}{The log2 dilution stepwise to assume for the titer
data, for example 1 if titer data relates to 2-fold dilutions or 0 if it is
continuous data.}
}
\value{
Returns a list of parameter estimates from the fit.

"slope_factors" is the factor by which fold-differences for each of the
antigens vs the homologous antigen vary according to the size of the
fold-differences seen in the reference_sr_group. For example a
sr_group_slope estimate of 0.5 says that fold-differences tended to be
about half on the log2 scale for that serum group, so a 4-fold drop for the
reference_sr_group would only be 2-fold for that group.

"ag_folddrops" Is a list of estimates for the baseline antigen
fold-differences seen across the serum groups (on the log2 scale). The
fold-difference for a particular serum group would be the ag folddrop *
serum group slope.

"logtiter_error_sigma" Is the estimate made for the standard deviation of
the fit error on the log2 scale.
}
\description{
This function takes an input of a table of data and different serum groups
and calculates how fold-drops against the different antigens measured in the
titer table relate to each other amongst the serum groups, relating
response breadth of each serum group by some factor to the response breadth in
the reference serum group. The typical use case is to compare the response
breadth of different vaccine groups, where the magnitude of fold-drops for
each of the antigens could be assumed to follow the same general pattern but
be reduced by some factor in some of the serum groups.
}
