% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmt.R
\name{estimate_sr_effects}
\alias{estimate_sr_effects}
\title{Estimate sera effects}
\usage{
estimate_sr_effects(
  titers,
  ci_method = "HDI",
  ci_level = 0.95,
  dilution_stepsize = 0,
  ag_mean_prior_mean = 0,
  ag_mean_prior_sigma = 1000,
  sigma_prior_alpha = 2,
  sigma_prior_beta = 5,
  sigma_prior_sr_effect = 4,
  options = list()
)
}
\arguments{
\item{titers}{A matrix of titers}

\item{ci_method}{The method to use when calculating the confidence intervals}

\item{ci_level}{The confidence level to use when calculating confidence intervals}

\item{dilution_stepsize}{The dilution stepsize. Generally 1 for titers of
the form '20', '40', '80', and 0 for titers that are read out continuously.}

\item{ag_mean_prior_mean}{Prior for the mean parameter of the antigen mean values}

\item{ag_mean_prior_sigma}{Prior for the standard deviation parameter of the antigen mean values}

\item{sigma_prior_alpha}{Prior for the alpha parameter of the inverse gamma distribution to estimate the standard deviation}

\item{sigma_prior_beta}{Prior for the beta parameter of the inverse gamma distribution to estimate the standard deviation}

\item{sigma_prior_sr_effect}{Prior for the serum effect}

\item{options}{Options for the sampler}
}
\description{
Estimate sera effects
}
