% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmt.R
\name{gmt_me}
\alias{gmt_me}
\title{Calculate the geometric mean titer of a set of titers}
\usage{
gmt_me(titers, ci_method = "quap", ci_level = 0.95, options = list())
}
\arguments{
\item{titers}{A matrix of titers}

\item{ci_method}{The method to use when calculating the confidence intervals}

\item{ci_level}{The confidence level to use when calculating confidence intervals}

\item{options}{Options for the sampler}
}
\description{
Calculate the geometric mean titer of a set of titers
}
