% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_gmt_titers}
\alias{impute_gmt_titers}
\title{Impute censored titers}
\usage{
impute_gmt_titers(result, titers)
}
\arguments{
\item{result}{Results from a titer GMT estimate from \code{impute_gmt_titers()}}

\item{titers}{A vector of titers for which to impute censored cases}
}
\value{
Returns a vector of titers with imputed values in place of censored
cases.
}
\description{
This function takes a vector of titers and imputes censored titers based on
the results of a geometric mean titer estimate from the \code{impute_gmt_titers()}
function. Censored titers are drawn from a censored normal distribution with
mean and standard deviation parameters equal to those provided in the
result argument.
}
