% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldchange.R
\name{log2diff_me}
\alias{log2diff_me}
\title{Calculate the mean difference between two paired sets of titers}
\usage{
log2diff_me(
  titers1,
  titers2,
  ci_method = "quap",
  ci_level = 0.95,
  options = list()
)
}
\arguments{
\item{titers1}{The first titerset from which to calculate the mean difference}

\item{titers2}{The second titerset to which to calculate the mean difference}

\item{ci_method}{The method to use when calculating the confidence intervals}

\item{ci_level}{The confidence level to use when calculating confidence intervals}

\item{options}{Options for the sampler}
}
\description{
This function is useful for example for calculating differences where you have a set of
pre-vaccination and post-vaccination samples and you would like to know the mean response
size, accounting for non-detectable values.
}
