% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titertools-package.R
\docType{package}
\name{titertools-package}
\alias{titertools-package}
\alias{titertools}
\title{The 'titertools' package.}
\description{
Provides a set of functions for statistical analysis of censored
data, in particular data such as titration data where there is typically
a detection threshold such as <10 titers.
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
